V2 Documentation updates (#7)
* drop scrolling-touch and scrolling-auto utilities

* add text-overflow docs

* move truncate from word-break docs to text-overflow docs

* add 2xl breakpoint documentation

* update docs for default line heights

* add transform docs

* update font-family examples to reflect the real values

* update docs about configuring variants

I dropped the docs about the before/after/without/variants utility functions. We can re-add them, but I thought this would be better if this was sort of deprecated (still usable) and only documented in the v1 docs.

* add some information about the max-w-prose utility

* rename flex-no-wrap to flex-nowrap

* rename whitespace-no-wrap to whitespace-nowrap

* update preflight docs

* add draft on presets

* replace default with DEFAULT

Skipping the cursor examples, because we do want cursor-default there

* make @apply great again

* add information about deeply nested color objects

* add documentation about the theme function resolving complex values

* add docs about the deep merging of arrays of objects

Not 100% sure if this needs to be documented here. Ref: https://github.com/tailwindlabs/tailwindcss/pull/2700

* update docs to reflect config file changes

default -> DEFAULT was already changed and written about in the upgrade guide

* wip: focus rings

* update color palettes

Co-authored-by: Brad Cornes <bradlc41@gmail.com>